A la fin d'une partie, sauvegarde la partie en base de données

+++ backend/src/main/java/cinequiz/backend/model/GameQuestion.java
@@ -19,5 +19,5 @@ public class GameQuestion {
     String playerAnswer;
 
     @ApiModelProperty("Time the player take to answer")
-    int answerTime;
+    float answerTime;
 }
